Revert "Present onboarding in a separate window on macOS (#3987)"
This reverts commit 6ad973b8b7c4c79346279ac6d663009f85cc91d2.
