+++ Sources/App/Onboarding/Steps/Servers/OnboardingServersListView.swift
@@ -9,7 +9,6 @@ struct OnboardingServersListView: View {
         static let delayUntilAutoconnect: TimeInterval = 2
     }
 
-    @EnvironmentObject var viewControllerProvider: ViewControllerProvider
     @Environment(\.dismiss) private var dismiss
     @Environment(\.horizontalSizeClass) private var sizeClass
 
@@ -21,16 +20,6 @@ struct OnboardingServersListView: View {
     @State private var autoConnectWorkItem: DispatchWorkItem?
     @State private var autoConnectInstance: DiscoveredHomeAssistant?
     @State private var autoConnectBottomSheetState: AppleLikeBottomSheetViewState?
-
-    private var presentingViewController: UIViewController {
-        if let providedController = viewControllerProvider.viewController, Current.isCatalyst {
-            return providedController
-        } else if let topController = UIApplication.shared.topViewController() {
-            return topController
-        } else {
-            fatalError("No controller provided for onboarding")
-        }
-    }
 
     let prefillURL: URL?
 
@@ -87,7 +76,7 @@ struct OnboardingServersListView: View {
         .sheet(isPresented: $showManualInput) {
             ManualURLEntryView { connectURL in
                 viewModel.manualInputLoading = true
-                viewModel.selectInstance(.init(manualURL: connectURL), presentingController: presentingViewController)
+                viewModel.selectInstance(.init(manualURL: connectURL))
             }
         }
         .fullScreenCover(isPresented: .init(get: {
@@ -142,7 +131,7 @@ struct OnboardingServersListView: View {
             Button {
                 autoConnectInstance = nil
                 guard let instance else { return }
-                viewModel.selectInstance(instance, presentingController: presentingViewController)
+                viewModel.selectInstance(instance)
             } label: {
                 Text(L10n.Onboarding.Servers.AutoConnect.button)
             }
@@ -187,17 +176,12 @@ struct OnboardingServersListView: View {
                     .progressViewStyle(.circular)
             } else {
                 Button(action: {
-                    if Current.isCatalyst {
-                        URLOpener.shared.open(
-                            AppConstants.WebURLs.homeAssistantGetStarted,
-                            options: [:],
-                            completionHandler: nil
-                        )
-                    } else {
-                        showDocumentation = true
-                    }
+                    showDocumentation = true
                 }, label: {
-                    Image(systemSymbol: .questionmark)
+                    Image(uiImage: MaterialDesignIcons.helpCircleOutlineIcon.image(
+                        ofSize: .init(width: 25, height: 25),
+                        color: .accent
+                    ))
                 })
                 .sheet(isPresented: $showDocumentation) {
                     SafariWebView(url: AppConstants.WebURLs.homeAssistantCompanionGetStarted)
@@ -234,7 +218,7 @@ struct OnboardingServersListView: View {
         if !screenLoaded {
             screenLoaded = true
             if let prefillURL {
-                viewModel.selectInstance(.init(manualURL: prefillURL), presentingController: presentingViewController)
+                viewModel.selectInstance(.init(manualURL: prefillURL))
             } else {
                 viewModel.startDiscovery()
             }
@@ -274,7 +258,7 @@ struct OnboardingServersListView: View {
             hideCenterLoader()
         }
         Button {
-            viewModel.selectInstance(.init(manualURL: url), presentingController: presentingViewController)
+            viewModel.selectInstance(.init(manualURL: url))
             viewModel.invitationLoading = true
         } label: {
             ZStack {
@@ -335,7 +319,7 @@ struct OnboardingServersListView: View {
 
     private func serverRow(instance: DiscoveredHomeAssistant) -> some View {
         Button(action: {
-            viewModel.selectInstance(instance, presentingController: presentingViewController)
+            viewModel.selectInstance(instance)
         }, label: {
             OnboardingScanningInstanceRow(
                 name: instance.bonjourName ?? instance.locationName,
@@ -387,27 +371,5 @@ struct OnboardingServersListView: View {
 #Preview {
     NavigationView {
         OnboardingServersListView(prefillURL: nil)
-    }
-}
-
-private extension UIApplication {
-    func topViewController(controller: UIViewController? = nil) -> UIViewController? {
-        let controller = controller ?? connectedScenes
-            .compactMap { $0 as? UIWindowScene }
-            .flatMap(\.windows)
-            .first(where: \.isKeyWindow)?
-            .rootViewController
-
-        if let navigationController = controller as? UINavigationController {
-            return topViewController(controller: navigationController.visibleViewController)
-        }
-        if let tabController = controller as? UITabBarController,
-           let selected = tabController.selectedViewController {
-            return topViewController(controller: selected)
-        }
-        if let presented = controller?.presentedViewController {
-            return topViewController(controller: presented)
-        }
-        return controller
     }
 }
